test: couvre le moteur clinique avec Vitest
- 31 tests sur detectRedFlags, interpretScore, generateHypothesis,
  PatientParser, extractMetricsFromTxt et l'aplatissement des données
- Test de non-régression sur le faux positif de drapeau rouge : un nom de
  champ contenant un terme d'alerte ne doit rien déclencher
- Vérifie qu'un score absent reste distinct d'un score à zéro, et que le
  contrôle qualité est bien recalculé plutôt que lu dans le fichier
- parseScore passe de parseInt à parseFloat : un score transmis en texte est
  désormais interprété comme le même score transmis en nombre
- Scripts pnpm test et test:watch

+++ utils/calculations.ts
@@ -4,7 +4,7 @@ import { SCORE_DEFINITIONS, RED_FLAGS, RedFlagDefinition } from "./definitions";
 const parseScore = (val: string | number | undefined | null): number | null => {
   if (val === undefined || val === null) return null;
   if (typeof val === "number") return val;
-  const parsed = parseInt(val);
+  const parsed = parseFloat(val);
   return isNaN(parsed) ? null : parsed;
 };
